Consolidate tree view error decoration providers
